+++ test/integration/spec/remotetracks.js
@@ -289,7 +289,7 @@ describe('RemoteVideoTrack', function() {
         testOptions: {
           bandwidthProfile: {
             video: {
-              maxTracks: 1,
+              maxSwitchedOnTracks: 1,
               dominantSpeakerPriority: 'low'
             }
           },