+++ test/integration/spec/room.js
@@ -604,7 +604,7 @@ describe('Room', function() {
     before(async () => {
       let thoseRooms;
       [, aliceRoom, thoseRooms] = await setup({
-        testOptions: { tracks: [], bandwidthProfile: { video: { maxTracks: 1 } } },
+        testOptions: { tracks: [], bandwidthProfile: { video: { maxSwitchedOnTracks: 1 } } },
         otherOptions: { tracks: [] },
         participantNames: ['Alice', 'Bob', 'Charlie'],
         nTracks: 0