+++ tsdef/AudioTrack.d.ts
@@ -4,7 +4,7 @@ export class AudioTrack extends Track {
   isStarted: boolean;
   isEnabled: boolean;
   kind: 'audio';
-  mediaStreamTrack: MediaStreamTrack;
+  mediaStreamTrack: MediaStreamTrack | null;
 
   attach(element?: HTMLMediaElement | string): HTMLMediaElement;
   detach(element?: HTMLMediaElement | string): HTMLMediaElement[];